[EHUFA-986] Fix failing test cases for Samsung tablet(tv-dpi) device.

+++ Source/UIKit/uit/src/androidTest/java/com/philips/platform/uit/matcher/TextViewPropertiesMatchers.java
@@ -40,7 +40,7 @@ public class TextViewPropertiesMatchers {
         };
     }
 
-    public static Matcher<View> isSameFontSize(final int expectedValue) {
+    public static Matcher<View> isSameFontSize(final float expectedValue) {
         return new BaseTypeSafteyMatcher<View>() {
             @Override
             protected boolean matchesSafely(View view) {
